requirements.txt
accelerate
aiohttp
fastapi
gpt4all
protobuf<=3.20
python-decouple
python-dotenv
sentence-transformers
sentry_sdk
uvicorn
